+++ lib/WindowLed/effects/fade.h
@@ -9,7 +9,7 @@ namespace Nezumikun {
     class EffectFade: public EffectWithTime {
       protected:
       public:
-        EffectFade(Canvas & canvas, uint8_t framePerSecond, uint16_t timeInMilliseconds = 1000) : EffectWithTime (canvas, framePerSecond, timeInMilliseconds) {
+        EffectFade(Canvas & canvas, uint8_t framePerSecond, uint32_t timeInMilliseconds = 1000) : EffectWithTime (canvas, framePerSecond, timeInMilliseconds) {
         };
         virtual void loop() override;
     };
