+++ lib/WindowLed/effects/with_time.cpp
@@ -4,7 +4,7 @@
 namespace Nezumikun {
   namespace WindowLed {
 
-    EffectWithTime::EffectWithTime(Canvas & canvas, uint8_t framePerSecond, uint16_t timeInMilliseconds) : Effect(canvas, framePerSecond) {
+    EffectWithTime::EffectWithTime(Canvas & canvas, uint8_t framePerSecond, uint32_t timeInMilliseconds) : Effect(canvas, framePerSecond) {
       this->timeInMilliseconds = timeInMilliseconds;
       this->frame = 0;
     };
